fix(DTFS2-7852): updated unit test cases
